2nd review: handling complex states
