+++ src/index.css
@@ -1,32 +1,40 @@
-.gralContainer{
+.gralContainer {
     display: flex;
-    flex-direction: column;
+    justify-content: space-around;
+    padding: 20px;
+}
+.leftContainer, .rightContainer{
+    width: 50%;
+    display: flex;
+    flex-direction: row;
     justify-content: center;
-    align-items: center;
-    background-color: #66339969;
-    border: 3px solid purple;
-    border-radius: 10px;
-    padding: 10px;
 }
-.btnsContainer{
-    display: flex;
+.leftScore{
+    padding-right: 5px;
+    margin: 0;
 }
-.tittle{
-    font-family: Arial, Helvetica, sans-serif;
+.rightScore{
+    padding-left: 5px;
+    margin: 0;
 }
-.counter{
-    font-size: 3.5rem;
+.leftBtn{
+    background-color: #2ee5e4;
 }
-.even{
-    font-size: 1.5rem;
+.leftBtn:hover {
+    background-color: rgb(0, 119, 255);
+    color: white;
 }
-.increaseBtn, .resetBtn, .decreaseBtn{
+.rightBtn{
+    background-color: #ff2d9e;
+    color: white;
+}
+.rightBtn:hover {
+    background-color: rgb(98, 43, 170);
+}
+.leftBtn, .rightBtn{
+    border: 1px solid black;
     border-radius: 5px;
-    border-color: purple;
-    background-color: rgb(255, 184, 255);
-    font-weight: bolder;
-    font-size: 1.2rem;
-}
-.counter, .even, .increaseBtn, .resetBtn, .decreaseBtn{
-    color: purple;
+    width: fit-content;
+    height: fit-content;
+    font-size: 20px;
 }
