+++ src/oldIndex.css
@@ -1,0 +1,32 @@
+.gralContainer{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background-color: #66339969;
+    border: 3px solid purple;
+    border-radius: 10px;
+    padding: 10px;
+}
+.btnsContainer{
+    display: flex;
+}
+.tittle{
+    font-family: Arial, Helvetica, sans-serif;
+}
+.counter{
+    font-size: 3.5rem;
+}
+.even{
+    font-size: 1.5rem;
+}
+.increaseBtn, .resetBtn, .decreaseBtn{
+    border-radius: 5px;
+    border-color: purple;
+    background-color: rgb(255, 184, 255);
+    font-weight: bolder;
+    font-size: 1.2rem;
+}
+.counter, .even, .increaseBtn, .resetBtn, .decreaseBtn{
+    color: purple;
+}
